explore pandoc's pdf generation via roff
At least for Linux Debian and Ubuntu, the generation of .pdf with
pandoc and (pdf)roff of groff may represent an alternative to
pdfLaTeX:  easy access/lightweight installation, single compilation
to yield .pdf with searchable text layer, and syntax highlighting
for many languages (including markdown and LaTeX so far excluded).

This commit is about a proof of concept intentionally constrained
to a few individual .md (and subsequent .pdf) files only.